Remove some unused code and add basic model tests
